feat(mcp): fold the two context reads into get_entity_details and merge the company-context pair
