--- openmetadata-mcp/src/main/java/org/openmetadata/mcp/tools/GetCompanyContextTool.java
@@ -1,109 +1,0 @@
-package org.openmetadata.mcp.tools;
-
-import static org.openmetadata.schema.type.MetadataOperation.VIEW_ALL;
-
-import java.io.IOException;
-import java.util.HashMap;
-import java.util.Map;
-import lombok.extern.slf4j.Slf4j;
-import org.openmetadata.schema.entity.context.ContextMemory;
-import org.openmetadata.schema.entity.context.ContextMemorySourceType;
-import org.openmetadata.schema.entity.context.MemoryVisibility;
-import org.openmetadata.service.Entity;
-import org.openmetadata.service.exception.EntityNotFoundException;
-import org.openmetadata.service.limits.Limits;
-import org.openmetadata.service.security.Authorizer;
-import org.openmetadata.service.security.auth.CatalogSecurityContext;
-import org.openmetadata.service.security.policyevaluator.OperationContext;
-import org.openmetadata.service.security.policyevaluator.ResourceContext;
-import org.openmetadata.service.util.FullyQualifiedName;
-
-/** Fetches a single Company Context knowledge pill (a {@link ContextMemory}) by FQN. */
-@Slf4j
-public class GetCompanyContextTool implements McpTool {
-
-  private static final String NOT_A_SHARED_PILL_ERROR =
-      "Requested entity is not a shared Company Context knowledge pill";
-
-  @Override
-  public Map<String, Object> execute(
-      Authorizer authorizer, CatalogSecurityContext securityContext, Map<String, Object> params)
-      throws IOException {
-    authorizer.authorize(
-        securityContext,
-        new OperationContext(Entity.CONTEXT_MEMORY, VIEW_ALL),
-        new ResourceContext<>(Entity.CONTEXT_MEMORY));
-    Map<String, Object> result;
-    String fqn = (String) params.get("fqn");
-    if (fqn == null || fqn.isBlank()) {
-      result = Map.of("error", "'fqn' parameter is required");
-    } else {
-      result = lookupPill(fqn);
-    }
-    return result;
-  }
-
-  private static Map<String, Object> lookupPill(String fqn) {
-    Map<String, Object> result;
-    try {
-      ContextMemory memory = fetchPill(fqn);
-      result =
-          isExposablePill(memory) ? projectPill(memory) : Map.of("error", NOT_A_SHARED_PILL_ERROR);
-    } catch (EntityNotFoundException e) {
-      result = Map.of("error", "No Company Context knowledge pill found for '" + fqn + "'");
-    }
-    return result;
-  }
-
-  /**
-   * A {@link ContextMemory} FQN is a single name part, so a file-extracted pill name (e.g. {@code
-   * report.md_<hash>}) carries dots and is stored quoted ({@code "report.md_<hash>"}). MCP clients
-   * routinely hand the value back unquoted; {@link FullyQualifiedName#quoteName(String)} restores
-   * the canonical quoting (and is a no-op when the value is already correctly quoted), so the
-   * by-name lookup resolves whichever form the client supplies.
-   */
-  private static ContextMemory fetchPill(String fqn) {
-    String normalizedFqn = FullyQualifiedName.quoteName(fqn);
-    LOG.debug("Getting company context pill: {} (normalized fqn: {})", fqn, normalizedFqn);
-    return Entity.getEntityByName(
-        Entity.CONTEXT_MEMORY, normalizedFqn, "sourceFile,owners,tags,domains", null);
-  }
-
-  /** Mirrors the search tool's scope: file-extracted pills with Shared visibility only. */
-  private static boolean isExposablePill(ContextMemory memory) {
-    return memory.getSourceType() == ContextMemorySourceType.FILE_EXTRACTION
-        && memory.getShareConfig() != null
-        && memory.getShareConfig().getVisibility() == MemoryVisibility.SHARED;
-  }
-
-  static Map<String, Object> projectPill(ContextMemory memory) {
-    Map<String, Object> pill = new HashMap<>();
-    pill.put("fullyQualifiedName", memory.getFullyQualifiedName());
-    pill.put("name", memory.getName());
-    putIfPresent(pill, "title", memory.getTitle());
-    putIfPresent(pill, "question", memory.getQuestion());
-    putIfPresent(pill, "answer", memory.getAnswer());
-    putIfPresent(pill, "summary", memory.getSummary());
-    putIfPresent(pill, "memoryType", memory.getMemoryType());
-    if (memory.getSourceFile() != null) {
-      pill.put("sourceFile", memory.getSourceFile().getFullyQualifiedName());
-    }
-    return pill;
-  }
-
-  private static void putIfPresent(Map<String, Object> pill, String key, Object value) {
-    if (value != null) {
-      pill.put(key, value);
-    }
-  }
-
-  @Override
-  public Map<String, Object> execute(
-      Authorizer authorizer,
-      Limits limits,
-      CatalogSecurityContext securityContext,
-      Map<String, Object> params) {
-    throw new UnsupportedOperationException(
-        "GetCompanyContextTool does not require limit validation.");
-  }
-}
